move localstorage helper in own module to make place for other root level mixins
